moved tx clear balance
