+++ x/stakeibc/client/cli/tx.go
@@ -599,6 +599,42 @@ func CmdCalibrateDelegation() *cobra.Command {
 	return cmd
 }
 
+func CmdClearBalance() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "clear-balance [chain-id] [amount] [channel-id]",
+		Short: "Broadcast message clear-balance",
+		Args:  cobra.ExactArgs(3),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argChainId := args[0]
+			argAmount, found := sdk.NewIntFromString(args[1])
+			if !found {
+				return errorsmod.Wrap(sdkerrors.ErrInvalidType, "can not convert string to int")
+			}
+			argChannelId := args[2]
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgClearBalance(
+				clientCtx.GetFromAddress().String(),
+				argChainId,
+				argAmount,
+				argChannelId,
+			)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
 func CmdUpdateInnerRedemptionRateBounds() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "set-redemption-rate-bounds [chainid] [min-bound] [max-bound]",
